requirements.txt
agent-starter-pack==0.2.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alabaster==1.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.2
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
chess==1.11.2
claude==0.4.11
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.5.0
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
dataclasses-json==0.6.7
datasets==4.5.0
debugpy==1.8.13
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
Django==5.1.7
django-allauth==65.6.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.22.4
executing==2.2.0
faiss-cpu==1.13.2
fastapi==0.126.0
fastjsonschema==2.21.1
ffmpy==1.0.0
filelock==3.18.0
Flask==3.1.3
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.5.1
gensim==4.4.0
google-api-core==2.24.2
google-auth==2.49.0
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.66.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gradio==6.9.0
gradio_client==2.3.0
greenlet==3.3.2
groovy==0.1.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.1
huggingface_hub==1.4.1
idna==3.10
imagesize==2.0.0
indic-nlp-library-itt==0.1.1
indictranstoolkit @ git+https://github.com/VarunGumma/IndicTransToolkit.git@3efb8418d0721b4ce267c2b3586899d313191357
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==1.2.10
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.17
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.14
llvmlite==0.44.0
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mcp==1.12.3
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.3
notebook==7.3.3
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.3
openai==1.99.1
openai-whisper==20250625
openfile==0.0.7
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
pypdf==6.7.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytube==15.0.0
pytz==2025.1
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roman-numerals==4.1.0
rpds-py==0.23.1
rsa==4.9
sacrebleu==2.6.0
sacremoses==0.1.1
safehttpx==0.1.7
safetensors==0.7.0
sarvamai==0.1.27
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==5.2.3
sentencepiece==0.2.1
setuptools==77.0.3
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.6
Sphinx==9.1.0
sphinx-argparse==0.5.2
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.48
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.1
tabulate==0.10.0
tenacity==9.1.4
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.22.2
tomlkit==0.13.3
toolwrapper==2.1.0
torch==2.6.0+cu124
torch-geometric==2.7.0
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.3.0
twilio==9.10.2
typer==0.24.0
typer-slim==0.24.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uctools==1.3.0
uri-template==1.3.0
urllib3==2.3.0
uuid_utils==0.14.1
uv==0.6.12
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==16.0
Werkzeug==3.1.6
wrapt==2.0.1
xgboost==2.1.4
xxhash==3.6.0
yarl==1.22.0
yts-downloader==0.1.2
zstandard==0.25.0
